flask-server/requirements.txt
Flask
requests
Pillow
opencv-python
scikit-image
